Merge the dockerfiles into one file
The dockerfiles were quite the same, except for the lists of package.
This situation made it difficult to maintain, especially when applying the
same change to every dockerfile.

To fix that, the dockerfiles have been merged into a a unique template.
When calling `make`, the dockerfile is written filling the template with the
given parameters.

Also, added settings for AT 13.0 and AT 14.0.

Signed-off-by: Erwan Prioul <erwan@linux.ibm.com>